rename impl feature to http-client
This also specifies that certain examples require the http-client.

As the `api_trait_implementation` example only requires serde_json
it is added as a dev-dependency.

Versions specified in Cargo.toml are using semver to install the latest.
Specifying 2.3 will install 2.6 which is counterintuitive.
Removing the minor version is more expressive about what is happening.