feat!: name clients Bot to differentiate from TelegramApi
With having both named `Api` and `TelegramApi` its not as easy to understand
what is what. One describes the interface (`TelegramApi`) and one is used
to actually interact with the Bot (`Bot`).

As the update to 0.39 will require users to explictly specify the new
features they will know the name of the http client they want anyway.