ci(rust): dont check features on beta
Running on beta mainly uses up a lot of resources without much benefit.
Doing this on stable and the MSRV is way more useful in comparison.
The main benefit of beta is to check that the beta channel works in
order to provide feedback to Rust. Thats already done with test and
clippy (and therefore check).